requirements.txt
bcrypt
blinker
certifi
click
cloudinary
colorama
Flask
Flask-Bcrypt
Flask-JWT-Extended
gunicorn
itsdangerous
Jinja2
MarkupSafe
mysql-connector
mysql-connector-python
packaging
passlib
protobuf
PyJWT
python-dotenv
six
urllib3
Werkzeug
